slashing and stabbing
