+++ src/ecs/components/LockPick.component.js
@@ -21,31 +21,4 @@ export default class LockPick extends Component {
       });
     }
   }
-
-  onTryStab(evt) {
-    const { interactee, interactor } = evt.data;
-    const damageBucket = this.entity.fireEvent("get-damage-types", {
-      damageTypes: [],
-    });
-
-    interactee.fireEvent("ApplyDamage", {
-      interactor,
-      interactee,
-      interactant: this.entity,
-      damageTypes: damageBucket.data.damageTypes,
-      verb: "stabs",
-    });
-  }
-
-  onGetMeleeInteractions(evt) {
-    const { interactions, interactee, interactor } = evt.data;
-    interactions.push({
-      name: "Stab",
-      evt: "try-stab",
-      interactor,
-      interactee,
-      interactant: this.entity,
-      caller: "interactant",
-    });
-  }
 }
