+++ src/ecs/index.js
@@ -28,6 +28,8 @@ import LockComponent from "./components/Lock.component";
 import LockPickComponent from "./components/LockPick.component";
 import LootComponent from "./components/Loot.component";
 import LuxComponent from "./components/Lux.component";
+import MeleeSlash from "./components/MeleeSlash.component";
+import MeleeStab from "./components/MeleeStab.component";
 import MotorComponent from "./components/Motor.component";
 import PCComponent from "./components/PC.component";
 import PositionComponent from "./components/Position.component";
@@ -77,6 +79,8 @@ engine.registerComponent(LockComponent);
 engine.registerComponent(LockPickComponent);
 engine.registerComponent(LootComponent);
 engine.registerComponent(LuxComponent);
+engine.registerComponent(MeleeSlash);
+engine.registerComponent(MeleeStab);
 engine.registerComponent(MotorComponent);
 engine.registerComponent(PCComponent);
 engine.registerComponent(PositionComponent);
